feature/sprint1deliverables (#142)
* Feature/skeleton UI init (#67)

* feat(init): initialized skeleton ui

* removed speed insight

it was erroring out and isn't worth fixing for now

* feat(init): setup custom theme

Also added 2 default themes

* no longer using cssnano

* feat: setup purgeCss to remove unused styles

* html brackets on the same line

* feat(init): created basic project layout with skeletonUI

resolves #27

* feat: added modals, toasts, and dark mode switch

* feat: setup modal registry for custom modal components

* feat: added transmission icon

* Create draft of database schema

Created a draft of the database schema for discussion on Friday.

Have not yet created a migration as changes yet to be discussed. There are some comments for discussion that should be removed before merging.

* Update README.md

* added husky install to README

* Add bcrypt to package.json

This commit adds the ‘bcrypt’ dependency to package.json. This dependency is required to generate and validate password hashes, which appear in the forthcoming seed data.

* Refactor database schema from first draftm

This commit makes the following refactors from the first draft of the schema:

- 'photo_url' is now 'photoUrl' - a bit of pythonism snuck in
- branches have a 'timezone' property which is an IANA timezone.
- 'equipment' is now 'accessory'
- lots of whitespace changes due to prisma format's blast radius

* Create seed data for initial schema

This commit adds some more expansive seed data for the schema defined in this branch.

Per @NathanGrenier I got rid of the original migration - not sure if it'll be neccesary to drop the database at Vercel so that it works.

* Revert "Create seed data for initial schema"

This reverts commit 0533ca7b9e0d36dce9e35f1fe010dbb975ba627d.

* refactor: re-added seed data

* fix: removed redundant migration and added basic schema migration

* fix: added secret corresponding to db name

* docs: create Sprint 1 folder, Add meeting notes

This commit adds the sprint 1 document deliverable folder, and adds notes about the meetings that
SvelteState has held so far.

* docs(file extension change): change file extensions of meetings notes

This commit changes the file extensions of the meeting notes to Markdown (md), which can be directly
viewed in GitHub

* test: fix workflow migration

* removed cat .env before initialization

* reverted changes

* docs: restructure Sprint Document Deliverable Directory

This commit restructures the directory in which SvelteState will be placing document deliverables
for each sprint

* docs: update README.md

This commit updates the README.md of the develop branch to reflect the one that is in the main
branch.

* docs: fix lint error

This commit fixes a lint error found in the README.md file

* docs: add Project Approach document deliverable

This commit adds the Project Approach and Technology Stack to the Sprint 1 document deliverable
folder

* docs: add Activity Log sheet to Sprint 1 deliverables folder

This commit adds the activity log sheet to the Sprint 1 deliverables folder

* ci: added release action

* revert: removed realease action. should go directly to main

---------

Co-authored-by: NathanGrenier <56941028+NathanGrenier@users.noreply.github.com>
Co-authored-by: NathanGrenier <nathangrenier01@gmail.com>
Co-authored-by: Brian Tkatch <brian@briantkatch.com>