Add session generation, verification and refresh
src/hooks.server.ts - checks session token on each payload. If valid, saves the critical user properties to the event locals. Critically, routes that are supposed to be open need to be whitelisted in this file.

src/app.d.ts: This contains the typescript definitions for the user event local.

src/lib/server/session.ts - this code generates and verifies the JWTs. There is some cleanup to be done here with TypeScript optionals.

Package file: had to add jsonwebtoken, and types for jsonwebtoken/bcrypt