Log in user after authentication
This commit amends the registration flow to log in the user and issue them a session after registration, and to redirect them to the ?destination query parameter.

It also passes through the ?destination parameter from the login page, if one is set there.

It would be nice to get the toast working again for account creation, but I'm not sure the best way to attack that.