Add UI for check-in route
This commit adds the UI for the check-in route. The UI is essentially
done, but the form action has not been written yet. Key features:

- Validation of the licence number in supported provinces AB, BC, ON, QC
- Capturing customer consent to the rental agreement

Still need to do the form action so that the check-in can be recorded in
the DB. There is some un-committed code on my machine that extends the
Dashboard to the CSR and provides a Check-in button for them; however,
that still needs to be worked on before a final commit.