fix(component): added blobCrud.svlete, a component for blobs / image handling
Currently, I need to implement two things: 1. Fix the creation of entries in all cases. I get a
hidden error. 2. I need to finish implementing blobCrud.svelte, which is a component of
car-form.svelte. This is a component made for a modal. The blobCrud.svelte is a component but has
difficulty connecting to +pages.server.svlete, to process the images to the server. Use blob-test to
see a functional example. I also need to cleanup the table that is generated after clicking an
entry. The modal replaces this table. Do not use the buttons that are created when the table is
generated.